+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 2.10.4
+- Fixed deprecated Git built-in extension API.
+
 ### 2.10.3
 - Fixed usage tracking function.
 
@@ -8,7 +11,7 @@
 - Fixed could not find usage tracking function.
 
 ### 2.10.0
-- Added usage tracking.
+- Added usage tracking function.
 - Amended `gitGrace.openWeb` command so it lists the current commit hash and branch names instead of the partial URLs.
 - Amended `gitGrace.sync` command so it does push before pull.
 - Amended `gitGrace.sync` command so it deletes merged local branches that have no remote counterparts.